Working on XMP Images
Images containing XMP format data will throw an uncatchable exception
due to incompatible version of xmpcore.jar version in Exifrotate.jar.
Since, we do not have access to Exifrotate.jar file's source, we wont be able
to fix this error or apply Error handling to it. This commit is part of
Ticket #729.